# apps/web/src/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "louped",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/launch": {
      "get": {
        "summary": "Launchables",
        "operationId": "launchables_api_launch_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Launchable"
                  },
                  "type": "array",
                  "title": "Response Launchables Api Launch Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Launch",
        "operationId": "launch_api_launch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LaunchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/launch/options": {
      "get": {
        "summary": "Launch Options",
        "operationId": "launch_options_api_launch_options_get",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Option"
                  },
                  "title": "Response Launch Options Api Launch Options Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/launch/jobs/{job_id}": {
      "get": {
        "summary": "Job",
        "operationId": "job_api_launch_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/launch/jobs/{job_id}/cancel": {
      "post": {
        "summary": "Cancel",
        "operationId": "cancel_api_launch_jobs__job_id__cancel_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Job": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "argv": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Argv"
          },
          "status": {
            "type": "string",
            "enum": ["queued", "running", "succeeded", "failed", "cancelled", "exported"],
            "title": "Status",
            "default": "queued"
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "title": "Created"
          },
          "started": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started"
          },
          "ended": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended"
          },
          "exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Code"
          }
        },
        "type": "object",
        "required": ["id", "title", "argv"],
        "title": "Job"
      },
      "JobDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "argv": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Argv"
          },
          "status": {
            "type": "string",
            "enum": ["queued", "running", "succeeded", "failed", "cancelled", "exported"],
            "title": "Status",
            "default": "queued"
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "title": "Created"
          },
          "started": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started"
          },
          "ended": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended"
          },
          "exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Code"
          },
          "log": {
            "type": "string",
            "title": "Log"
          }
        },
        "type": "object",
        "required": ["id", "title", "argv", "log"],
        "title": "JobDetail"
      },
...
      "LaunchRequest": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "options": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "boolean"
                },
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                }
              ]
            },
            "type": "object",
            "title": "Options",
            "default": {}
          },
          "config": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          },
          "recipe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipe"
          }
        },
        "type": "object",
        "required": ["id"],
        "title": "LaunchRequest"
      },
      "Launchable": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "group": {
            "type": "string",
            "title": "Group"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "config": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          },
          "recipe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipe"
          }
        },
        "type": "object",
        "required": ["id", "group", "title"],
        "title": "Launchable"
      },
...
      "Option": {
        "properties": {
          "flag": {
            "type": "string",
            "title": "Flag"
          },
          "kind": {
            "type": "string",
            "enum": ["text", "bool", "list", "choice"],
            "title": "Kind"
          },
          "default": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default"
          },
          "help": {
            "type": "string",
            "title": "Help",
            "default": ""
          },
          "choices": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Choices",
            "default": []
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": false
          }
        },
        "type": "object",
        "required": ["flag", "kind", "default"],
        "title": "Option",
        "description": "One flag of a command, as its parser declares it."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}